Move files from lib to src and index into src (#85)
* move lib to src

* path updates
- as suggested by @johnhooks to set as source for tests the "src" folder (and not the builded one)
- updated node internal modules paths prefix (ref. https://nodejs.org/api/module.html#modules-nodemodule-api)

* typescript file compilation config
