+++ .gitignore
@@ -1,4 +1,5 @@
 /node_modules
+/lib
 /@types
 npm-debug.log
 .DS_Store
