+++ package.json
@@ -18,15 +18,15 @@
     "node": ">=18"
   },
   "scripts": {
-    "lint": "eslint lib/*.js test/*.js index.js",
+    "lint": "eslint src/*.js test/*.js",
     "test-generate-mo": "msgfmt test/fixtures/latin13.po -o test/fixtures/latin13.mo & msgfmt test/fixtures/utf8.po -o test/fixtures/utf8.mo & msgfmt test/fixtures/obsolete.po -o test/fixtures/obsolete.mo",
     "test": "mocha",
     "preversion": "npm run lint && npm test",
     "postversion": "git push && git push --tags",
     "prepublish": "tsc && npm run lint && npm run test"
   },
-  "main": "./index.js",
-  "types": "./index.d.ts",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
   "license": "MIT",
   "dependencies": {
     "content-type": "^1.0.5",
@@ -48,7 +48,7 @@
     "eslint-plugin-n": "^16.6.2",
     "eslint-plugin-promise": "^6.1.1",
     "mocha": "^10.3.0",
-    "typescript": "^5.4.4"
+    "typescript": "^5.4.5"
   },
   "keywords": [
     "i18n",
