+++ test/shared.js
@@ -1,11 +1,9 @@
-'use strict';
-
+import { promisify } from 'node:util';
+import path from 'node:path';
+import { readFile as fsReadFile } from 'node:fs';
+import { fileURLToPath } from 'node:url';
 import * as chai from 'chai';
-import { promisify } from 'util';
-import path from 'path';
-import { formatCharset, parseHeader, generateHeader, foldLine, parseNPluralFromHeadersSafely } from '../lib/shared.js';
-import { readFile as fsReadFile } from 'fs';
-import { fileURLToPath } from 'url';
+import { formatCharset, parseHeader, generateHeader, foldLine, parseNPluralFromHeadersSafely } from '../src/shared.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
